Code updated so that bibretrieve can work zith ZBmath again.
As of 2015, it seems it is not possible anymore to query ZBmath and obtain
results directly in bibtex format. Hence it was not possible to use ZBmath
with the old version of bibretrieve.

In the new version, there are now two possibilities to define a backend: a
simplified one, as before, via bibretrieve-BACKEND-create-url, which just
builds a url for retrieving bibtex entries, and a more flexible one, via
a function named bibretrieve-BACKEND-http, which takes care of everything and
returns a buffer with bibtex entries.